[fix] JSH 오류 수정
