+++ src/Algorithm_Study/daily/CSY/D20250226.java
@@ -1,0 +1,4 @@
+package Algorithm_Study.daily.CSY;
+
+public class D20250226 {
+}
